requirements.txt
streamlit==1.32.0
mistralai==0.0.10
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
websockets==12.0
python-multipart==0.0.9
pydantic==2.6.3 